+++ lib/resque/data_store.rb
@@ -12,7 +12,6 @@ module Resque
       @failed_queue_access  = FailedQueueAccess.new(@redis)
       @workers              = Workers.new(@redis)
       @stats_access         = StatsAccess.new(@redis)
-      @redis_time_available = redis_time_available?
     end
 
     def_delegators :@queue_access, :push_to_queue,
@@ -92,14 +91,23 @@ module Resque
     end
 
     def server_time
-      time, _ = @redis_time_available ? @redis.time : Time.now
+      time, _ = redis_time_available? ? @redis.time : Time.now
       Time.at(time)
     end
 
+    # only needed until support for Redis 2.4 is removed
     def redis_time_available?
-      @redis.time
-    rescue Redis::CommandError
-      false
+      if @redis_time_available.nil? && !Resque.inline
+        begin
+          @redis_time_available = @redis.time
+        rescue Redis::CommandError
+          @redis_time_available = false
+        end
+      elsif Resque.inline
+        false
+      else
+        @redis_time_available
+      end
     end
     private :redis_time_available?